requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0

# OAuth & JWT
python-jose[cryptography]==3.3.0
httpx==0.28.1

# Development
python-dotenv==1.0.1
black==24.10.0
isort==5.13.2
ruff==0.8.6

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
